requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
httpx>=0.27.0
python-multipart>=0.0.9
gunicorn>=22.0.0
PyJWT[crypto]>=2.8.0
